WebSockets Next: merge BlockingSender and Sender interfaces
- fixes #44470
